Appointment Agent/Using Groq/requirements.txt
streamlit
groq
google-generativeai
python-dotenv

